Move deviceufunc into CUDA target
